A little bit of sugar added to clusto-shell:
This allows for basic "batch mode" in clusto shell, meaning:
 - you can pipe a script to clusto shell (cat script.py | clusto-shell)
 - you can redirect a script to clusto shell (clusto-shell < script.py)

My original goal was to be able to use clusto-shell as the interpreter
in a script (ala #!/usr/bin/env clusto-shell) but the IPython people
explicitly prohibit cmd line execution in embedded shells as stated
in IPython/ipmaker.py ~703:
"""
Do NOT execute files named in the command line as scripts to be loaded
by embedded instances.  Doing so has the potential for an infinite
recursion if there are exceptions thrown in the process.
"""
Bummer.

Some considerations when writing scripts to use in clusto-shell:
 - this behaves as normal ipython program, i.e. get_entities() will
   print the list even if no explicit `print` call is added
 - scripts must end with a newline, this becase ipython does "guessing"
   when writing code interactively. Think how you do for loops in
   ipython